\name{fpcond}
\alias{fpcond}
\title{fpcond}
\description{

  Calculeaza valoarea densitatii conditionata X|Y=y sau Y|X=x cu x si y dat.

}
\usage{fpcond(f,marginala,valX,valY)}

\arguments{

  \item{f}{Densitatea bidimensionala a lui X si Y}
  \intem{marginala}{Marginala celui care conditioneaza}
  \item{valY}{Valoarea lui y}
  \item{valX}{Valoarea lui x}
}

\value{Valoarea densitatii conditionata X|Y=y sau Y|X=x. }
