\name{fPvabidimcont}
\alias{fPvabidimcont}
\title{fPvabidimcont}
\description{

 Calculeaza P(lowX < X < highY, lowY< Y < highY).

}
\usage{fPvabidimcont <-  function(fdens,lowX=-10^5,highX=10^5,lowY=-10^5,highY=10^5)}

\arguments{

  \item{fdens}{Densitatea de probabilitate a v.a. bidimensionale}
  \item{lowX}{Lower bound-ul pentru parametrul x}
  \item{highX}{Upper bound-ul pentru parametrul x}
  \item{lowY}{Lower bound-ul pentru parametrul y}
  \item{highY}{Upper bound-ul pentru parametrul y}


}

\details{

}
\value{Valoarea probabilitatii. }
